tema 6

+++ desdoble/TEMA 5/ejercicio20/src/ejercicio20/Cuenta.java
@@ -21,7 +21,7 @@ public class Cuenta {
     }
     
     //Parametrizado
-    public Cuenta(int saldo) {
+    public Cuenta(double saldo) {
     
         this.saldo = saldo;
     }
